apps/workflow-engine/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
httpx
simpleeval
sqlmodel
sse-starlette
aiosqlite
asyncpg
psycopg2-binary
croniter
cryptography
pandas
openpyxl
google-genai
google-api-core
openai
anthropic
tiktoken>=0.5.0  # Token counting for memory subnodes
numpy>=1.24.0    # Vector operations for semantic memory
neo4j>=5.0.0     # Knowledge graph memory (Neo4j driver)
# aiosmtplib  # Uncomment if using SMTP for email
